backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["hiki7 <rsaliyev03@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
djangorestframework = "^3.15.2"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.5.0"
psycopg2 = "^2.9.10"
azure-communication-email = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
